back/requirements/test.txt
-r ./base.txt

django-environ==0.11.2
django-querycount==0.8.3
djhtml==3.0.6
freezegun==1.5.1
ruff==0.6.9
pytest==8.3.3
pytest-django==4.9.0
sqlfluff==3.1.1
requests-mock==1.12.1
